+++ CHANGELOG.rst
@@ -31,6 +31,9 @@ Added
 
 - ``fetch_schema`` option to ``cartridge.pool.connect``.
 
+- ``check_cookie_hash`` parameter in stateful failover configuration
+  (`#1765 <https://github.com/tarantool/cartridge/issues/1765>`_).
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -46,6 +49,9 @@ Fixed
 
 - Use default values of ``replication_connect_quorum`` and
   ``replication_sync_timeout`` in first ``box.cfg`` call.
+
+- Check hash of cluster cookie on stateful failover configuration
+  (`#1765 <https://github.com/tarantool/cartridge/issues/1765>`_).
 
 -------------------------------------------------------------------------------
 [2.7.7] - 2022-12-09